Incognito mode: (search and article) history on/off
Added a preference option to enable/disable both the
search and article history.

When disabling, the search and article history are
automatically deleted, and the search input does not
show previous search strings.

Info messages are shown on search and article history
if the histories are disabled in the settings.

Follow up to I6697eca28e6683101100361cf3dd95d4860d3154

Bug: T96546
Bug: T115776
Change-Id: I92ec34150a37e37de18770c52abf8fd266512ab8